Move global config into config.ts
