+++ weave-js/src/entrypoint.tsx
@@ -9,25 +9,7 @@ import PagePanel from './components/PagePanel';
 import {WeaveMessage} from './components/Panel2/WeaveMessage';
 import {NotebookComputeGraphContextProvider} from './contextProviders';
 import {URL_BROWSE, URL_LOCAL, URL_RECENT, URL_WANDB} from './urls';
-
-// These get populated via /__frontend/env.js and are defined in weave_server.py
-declare global {
-  interface Window {
-    WEAVE_CONFIG: {
-      PREFIX: string;
-      ANALYTICS_DISABLED: boolean;
-      WEAVE_BACKEND_HOST: string;
-    };
-  }
-}
-if (!window.WEAVE_CONFIG) {
-  console.warn('Unable to get configuration from server, using defaults');
-  window.WEAVE_CONFIG = {
-    PREFIX: '',
-    ANALYTICS_DISABLED: false,
-    WEAVE_BACKEND_HOST: '/__weave',
-  };
-}
+import getConfig from './config';
 
 class ErrorBoundary extends React.Component<{}, {hasError: boolean}> {
   static getDerivedStateFromError(error: Error) {
@@ -78,8 +60,9 @@ const Main = ({browserType}: MainProps) => (
   </React.Suspense>
 );
 
+const basename = getConfig().PREFIX;
 ReactDOM.render(
-  <Router basename={window.WEAVE_CONFIG.PREFIX}>
+  <Router basename={basename}>
     <Switch>
       <Route path={`/${URL_BROWSE}/${URL_RECENT}/:assetType?`}>
         <Main browserType={URL_RECENT} />
